+++ src/network/devices/linux/command-kernel/createLinuxHostShell.ts
@@ -21,6 +21,7 @@ import { HeadCommand } from './commands/Head';
 import { IdCommand } from './commands/Id';
 import { LnCommand } from './commands/Ln';
 import { LocaleCommand } from './commands/Locale';
+import { LoggerCommand } from './commands/Logger';
 import { LsCommand } from './commands/Ls';
 import { MkdirCommand } from './commands/Mkdir';
 import { MktempCommand } from './commands/Mktemp';
@@ -113,6 +114,7 @@ export function createLinuxHostShell(deps: LinuxMachineApiDeps): Interpreter {
   registry.register(() => new MktempCommand());
   registry.register(() => new FileCommand());
   registry.register(() => new LocaleCommand());
+  registry.register(() => new LoggerCommand());
   registry.register(() => new Md5sumCommand());
   registry.register(() => new Sha1sumCommand());
   registry.register(() => new Sha256sumCommand());